Fiks navnevisning
